send test

+++ trs.h
@@ -93,7 +93,7 @@ int trs_send(int destination_fd, unsigned char message_type, char* data, size_t 
 
     // Send it off.
     size_t total_sent = 0;
-    size_t sent;
+    int sent;
     size_t left;
     while (total_sent < total_len) {
         left = total_len - total_sent;
@@ -101,7 +101,7 @@ int trs_send(int destination_fd, unsigned char message_type, char* data, size_t 
         sent = send(destination_fd, &bufsend[total_sent], left, 0);
         if (sent > 0) {
             total_sent = total_sent + sent;
-            printf("Sent %zu out of %zu bytes\n", sent, total_len);
+            printf("Sent %d out of %zu bytes\n", sent, total_len);
         }
     }
